refactor(cli): split raw_mode module into file
